linguabot-chat-service/speech/requirements.txt
fastapi
google-cloud-speech
uvicorn
python-multipart